requirements.txt

pandas==2.2.3
numpy==2.1.2
pyyaml==6.0.2
jinja2==3.1.4
matplotlib==3.9.2
requests==2.32.3
